backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
groq==0.9.0
pymupdf==1.24.3
python-docx==1.1.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
pillow==10.3.0
aiofiles==23.2.1
email-validator==2.1.1
requests==2.31.0
